+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 - Add the `Password` dialog box.
+- Add the `Zenity` backend.
 
 # v0.1.0 (2019-01-08)
 - Initial release with the `Input`, `Message` and `Question` dialog boxes and